Refactor utility functions, including gfa construction
